Add a --dry-run flag to the release tool's prepare command to allow safe testing of the release pipeline, and refactor the tool's codebase into modular components to improve maintainability. This structure is verified by an updated test suite using temporary directories and package-level mocking.
